# src/tsoa-auto/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Project": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"start_date": {
						"type": "string"
					},
					"end_date": {
						"type": "string"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/Skill"
						},
						"type": "array"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"department": {
						"$ref": "#/components/schemas/Department"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Employee": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"date_of_joining": {
						"type": "string",
						"format": "date-time"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"department": {
						"$ref": "#/components/schemas/Department"
					},
					"branch": {
						"$ref": "#/components/schemas/Branch"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"buddies": {
						"items": {
							"$ref": "#/components/schemas/Buddies"
						},
						"type": "array"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/Skill"
						},
						"type": "array"
					},
					"leaverequests": {
						"items": {
							"$ref": "#/components/schemas/LeaveRequest"
						},
						"type": "array"
					},
					"hrletters": {
						"items": {
							"$ref": "#/components/schemas/HRLetters"
						},
						"type": "array"
					},
					"clockins": {
						"items": {
							"$ref": "#/components/schemas/Clockin"
						},
						"type": "array"
					},
					"project": {
						"$ref": "#/components/schemas/Project"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"role_name": {
						"type": "string"
					},
					"role_description": {
						"type": "string"
					},
					"employee": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/Permission"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Permission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"permission_name": {
						"type": "string"
					},
					"permission_description": {
						"type": "string"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/Role"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Department": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"branch": {
						"$ref": "#/components/schemas/Branch"
					},
					"projects": {
						"items": {
							"$ref": "#/components/schemas/Project"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Branch": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"departments": {
						"items": {
							"$ref": "#/components/schemas/Department"
						},
						"type": "array"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Company": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"logo_url": {
						"type": "string"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"branches": {
						"items": {
							"$ref": "#/components/schemas/Branch"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Buddies": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"buddy_group_name": {
						"type": "string"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"buddy_tasks": {
						"items": {
							"$ref": "#/components/schemas/BuddyTask"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"BuddyTask": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"buddy_task": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"buddies": {
						"$ref": "#/components/schemas/Buddies"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Skill": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"projects": {
						"items": {
							"$ref": "#/components/schemas/Project"
						},
						"type": "array"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LeaveRequest": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"reason": {
						"type": "string"
					},
					"from_date": {
						"type": "string"
					},
					"to_date": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"employee": {
						"$ref": "#/components/schemas/Employee"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"HRLetters": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"letter_subject": {
						"type": "string"
					},
					"letter_content": {
						"type": "string"
					},
					"letter_time": {
						"type": "string"
					},
					"employee": {
						"$ref": "#/components/schemas/Employee"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Clockin": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"clock_in": {
						"type": "string",
						"format": "date-time"
					},
					"clock_out": {
						"type": "string",
						"format": "date-time"
					},
					"employee": {
						"$ref": "#/components/schemas/Employee"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Task": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"project": {
						"$ref": "#/components/schemas/Project"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResCompany": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"logo_url": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResBranch": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"company": {
						"$ref": "#/components/schemas/ResCompany"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResDepartment": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"branch": {
						"$ref": "#/components/schemas/ResBranch"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResRole": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"roleName": {
						"type": "string"
					},
					"roleDescription": {
						"type": "string"
					},
					"permission": {
						"items": {
							"$ref": "#/components/schemas/ResPermission"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResPermission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"permission_name": {
						"type": "string"
					},
					"permission_description": {
						"type": "string"
					},
					"role": {
						"items": {
							"$ref": "#/components/schemas/ResRole"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResSkill": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResEmployee": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"username": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					},
					"department": {
						"$ref": "#/components/schemas/ResDepartment"
					},
					"branch": {
						"$ref": "#/components/schemas/ResBranch"
					},
					"company": {
						"$ref": "#/components/schemas/ResCompany"
					},
					"role": {
						"$ref": "#/components/schemas/ResRole"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/ResSkill"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResTask": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"project": {
						"$ref": "#/components/schemas/ResProject"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResProject": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"department": {
						"$ref": "#/components/schemas/ResDepartment"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/ResEmployee"
						},
						"type": "array"
					},
					"end_date": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/ResSkill"
						},
						"type": "array"
					},
					"start_date": {
						"type": "string"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/ResTask"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReqTask": {
				"properties": {
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					}
				},
				"required": [
					"name",
					"start_date",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResEmployeeN": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"username": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					},
					"department": {
						"$ref": "#/components/schemas/ResDepartment"
					},
					"branch": {
						"$ref": "#/components/schemas/ResBranch"
					},
					"company": {
						"$ref": "#/components/schemas/ResCompany"
					},
					"role": {
						"$ref": "#/components/schemas/ResRole"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResLeaveRequest": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"from_date": {
						"type": "string"
					},
					"to_date": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"reason": {
						"type": "string"
					},
					"employee": {
						"$ref": "#/components/schemas/ResEmployeeN"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResError": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReqLeaveRequest": {
				"properties": {
					"from_date": {
						"type": "string"
					},
					"to_date": {
						"type": "string"
					},
					"status": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"reason": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReqLeaveRequestN": {
				"properties": {
					"status": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Api-Token": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "WorkHive APIs",
		"version": "0.0.1",
		"description": "WorkhiveServer",
		"contact": {}
	},
	"paths": {
		"/project/{projectId}/task/{projectId}": {
			"get": {
				"operationId": "GetTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/project/{projectId}/task": {
			"post": {
				"operationId": "SaveTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResTask"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqTask"
							}
						}
					}
				}
			}
		},
...
		"/leaveRequest/{leaveRequestId}": {
			"get": {
				"operationId": "GetOneLeaveRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ResLeaveRequest"
										},
										{
											"$ref": "#/components/schemas/ResError"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Leave_Request"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "leaveRequestId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/leaveRequest/yourlr": {
			"post": {
				"operationId": "GetAllLeaveRequestSelf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ResLeaveRequest"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Leave_Request"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": []
			}
		},
		"/leaveRequest": {
			"get": {
				"operationId": "GetAllLeaveRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ResLeaveRequest"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Leave_Request"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "SaveLeaveRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResLeaveRequest"
								}
							}
						}
					}
				},
				"description": "saves leave requests",
				"summary": "saves leave requests",
				"tags": [
					"Leave_Request"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqLeaveRequest"
							}
						}
					}
				}
			}
		},
		"/leaveRequest/{leaverequestId}": {
			"put": {
				"operationId": "UpdateLeaveRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResLeaveRequest"
								}
							}
						}
					}
				},
				"description": "updates leave request",
				"summary": "updates leave request",
				"tags": [
					"Leave_Request"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "leaverequestId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqLeaveRequestN"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}